Fixes function lookup that needs implicit conversion of integral_constant for gcc-4.4
Doing make check with gcc-4.4 resulted in:

/home/mblatt/buildbot-slave1/DUNE-1/git/master/dune-geometry/dune/geometry/multilineargeometry.hh:227: error: no matching function for call to ‘Dune::GeometryType::GeometryType(std::integral_constant<unsigned int, 3u>, const int&)’
/home/mblatt/buildbot-slave1/DUNE-1/git/master/dune-geometry/dune/geometry/type.hh:135: note: candidates are: Dune::GeometryType::GeometryType(int)
/home/mblatt/buildbot-slave1/DUNE-1/git/master/dune-geometry/dune/geometry/type.hh:125: note:                 Dune::GeometryType::GeometryType(unsigned int)
/home/mblatt/buildbot-slave1/DUNE-1/git/master/dune-geometry/dune/geometry/type.hh:103: note:                 Dune::GeometryType::GeometryType(unsigned int, unsigned int)
/home/mblatt/buildbot-slave1/DUNE-1/git/master/dune-geometry/dune/geometry/type.hh:63: note:                 Dune::GeometryType::GeometryType(Dune::GeometryType::BasicType, unsigned int)
/home/mblatt/buildbot-slave1/DUNE-1/git/master/dune-geometry/dune/geometry/type.hh:58: note:                 Dune::GeometryType::GeometryType()
/home/mblatt/buildbot-slave1/DUNE-1/git/master/dune-geometry/dune/geometry/type.hh:25: note:                 Dune::GeometryType::GeometryType(const Dune::GeometryType&)

Aparrently gcc-4.4 has problems figuring out the needed implicit
cast. Therfore we explicitly cast to the underlying type with this commit.

+++ dune/geometry/multilineargeometry.hh
@@ -224,7 +224,7 @@ namespace Dune
     }
 
     /** \brief obtain the name of the reference element */
-    Dune::GeometryType type () const { return GeometryType( topologyId(), mydimension ); }
+    Dune::GeometryType type () const { return GeometryType( (unsigned int)topologyId(), mydimension ); }
 
     /** \brief obtain number of corners of the corresponding reference element */
     int corners () const { return refElement().size( mydimension ); }
